Refactor Folder Structure and Introduce Import Aliases Fixes #44
