+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as TodoList } from "./todo/TodoList";
+export { default as Counter } from "./counter/Counter";
