+++ vite.config.ts
@@ -7,4 +7,9 @@ export default defineConfig({
     host: "0.0.0.0",
     port: 3000,
   },
+  resolve: {
+    alias: {
+      "@": "/src",
+    },
+  },
 });
