+++ vitest.config.ts
@@ -1,8 +1,9 @@
 import { defineConfig } from "vitest/config";
 import react from "@vitejs/plugin-react";
+import viteConfig from "./vite.config";
 
 export default defineConfig({
-  plugins: [react()],
+  ...viteConfig,
   test: {
     globals: true,
     environment: "jsdom",
